requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
allure-pytest
allure-python-commons
annotated-types
apipkg
appdirs
astroid
async-generator
atomicwrites
attrs
autopep8
babel
backports.entry-points-selectable
beautifulsoup4
bidict
blinker
bs4
cachelib
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
configparser
contourpy
crayons
cycler
decorator
deepdiff
distlib
dnspython
docxcompose
docxtpl
exceptiongroup
execnet
faiss-cpu
Faker
filelock
Flask
Flask-Ext
Flask-Mail
flask-marshmallow
Flask-MySQL
Flask-MySQLdb
Flask-Session
Flask-SocketIO
Flask-SQLAlchemy
flask-swagger-ui
fonttools
frozenlist
ghp-import
google
google-genai
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
helium
httplib2
idna
importlib_metadata
iniconfig
isort
itsdangerous
Jinja2
jmespath
jsonpath
jsonpath-ng
jsonschema
jsonschema-specifications
kiwisolver
lazy-object-proxy
lxml
Markdown
MarkupSafe
marshmallow
matplotlib
mccabe
mergedeep
mkdocs
mkdocs-get-deps
MouseInfo
msgspec
multidict
mysql
mysql-connector-python
marshmallow-sqlalchemy
mysqlclient
numpy
oauthlib
openai
ordered-set
orderly-set
outcome
packaging
passlib
pathspec
pillow
pip
platformdirs
pluggy
ply
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
PyJWT
PyMySQL
pyparsing
pyperclip
pyrsistent
PySocks
pytest
python-dateutil
python-docx
python-dotenv
python-engineio
python-socketio
pytz
PyYAML
pyyaml_env_tag
referencing
requests
requests-oauthlib
rpds-py
rsa
selenium
setuptools
simple-websocket
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
scikit-learn
text-unidecode
tk
toml
tomli
tqdm
trio
trio-websocket
typed-ast
typing_extensions
typing-inspection
tzdata
uritemplate
urllib3
watchdog
webdriver-manager
websocket-client
Werkzeug
wget
wheel
wrapt
wsproto
WTForms
yarl
zipp
